This is a commit with Add Counter
Experiments: Using Verilog HDL, created 3 bit Add Counter. (When there
are POSEDGE SIGNAL detected by module and the variable en is 1, the reg
variable out_q will increase 1, when it is 111, it will become 000)

This experiment have implemented and indcated a way to deal with posedge
signal within a simulation file. In previous experiment, the code "top
-> clk = posedge" or "top -> clk = !clk" can not function, because 1)
posedge is not defined in scope; 2) clk have not been assigned with any
values so !clk is invalid. Solution: Using 2 timeInc, by implementing 2
seconds, one with high (1) and one with low (0), verilator can
understand the posedge signal and process/function as pretented.